# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Imagine API
  version: 0.1.0
  description: OpenAPI specification for the Imagine API (chi-based)
servers:
  - url: http://localhost:7770
    description: Local API
paths:
...
  /images/{uid}/download:
    get:
      summary: Create short-lived download token and redirect
      description: Creates a 5-minute download token and redirects to /images/{uid}/file?download=1&token=...
      operationId: quickDownloadImage
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
          description: Image UID
      responses:
        "302":
          description: Redirect to /images/{uid}/file with token
          headers:
            Location:
              schema:
                type: string
              description: Target URL with download=1 and token query params
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error (failed to create token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /download:
    post:
      summary: Download a set of images as a ZIP (requires token)
      description: Download images using a token from /download/sign. Supports password-protected tokens and embed controls.
      operationId: downloadImages
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
          description: Download token from /download/sign (required)
        - in: query
          name: password
          required: false
          schema:
            type: string
          description: Password if the token is password-protected (optional)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadRequest"
      responses:
        "200":
          description: ZIP archive
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (invalid/expired token or incorrect password)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (downloads not allowed or embedding not permitted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /download/sign:
    post:
      summary: Create a download token with optional access controls
      description: Generate a persistent download token for a set of images with configurable permissions and optional password protection
      operationId: signDownload
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignDownloadRequest"
      responses:
        "200":
          description: Download token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadToken"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/stats:
    get:
      summary: Get SSE connection statistics
      operationId: getSSEStats
      responses:
        "200":
          description: Connection statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSEStatsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/history:
    get:
      summary: Get recent event history
      operationId: getEventHistory
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of events to return
      responses:
        "200":
          description: Event history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventHistoryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Clear event history
      operationId: clearEventHistory
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/broadcast:
    post:
      summary: Broadcast event to all connected clients
      operationId: broadcastSSEEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SSEBroadcastRequest"
      responses:
        "200":
          description: Event broadcasted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSEBroadcastResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/send/{clientId}:
    post:
      summary: Send event to specific client
      operationId: sendToSSEClient
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SSEBroadcastRequest"
      responses:
        "200":
          description: Event sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  clientId:
                    type: string
                required: [success, message, clientId]
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SSEStatsResponse:
      type: object
      properties:
        connectedClients:
          type: integer
        clientIds:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
      required: [connectedClients, clientIds, timestamp]
...
    EventRecord:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        event:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [timestamp, event, data]
    EventHistoryResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventRecord"
        count:
          type: integer
      required: [events, count]
    SSEBroadcastRequest:
      type: object
      properties:
        event:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [event, data]
    SSEBroadcastResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        clients:
          type: integer
      required: [success, message, clients]
...
    DownloadRequest:
      type: object
      properties:
        uids:
          type: array
          items:
            type: string
        filename:
          type: string
      required: [uids]
    SignDownloadRequest:
      type: object
      description: Request to create a download token
      properties:
        uids:
          type: array
          items:
            type: string
          description: Array of image UIDs to include in the download token
        expires_in:
          type: integer
          description: Time in seconds until the token expires (0 for no expiry, default 900 = 15 minutes)
        allow_download:
          type: boolean
          description: Allow downloads using this token (default true)
          default: true
        allow_embed:
          type: boolean
          description: Allow embedding images on external sites (default false to prevent hotlinking)
          default: false
        show_metadata:
          type: boolean
          description: Include EXIF and other metadata in responses (default true)
          default: true
        password:
          type: string
          description: Optional password protection for the token (will be bcrypt hashed)
        description:
          type: string
          maxLength: 500
          description: Optional description of this share/download link
    DownloadToken:
      type: object
      description: Persistent download token with granular access controls
      properties:
        uid:
          type: string
          maxLength: 128
          description: 64-character hex token that serves as both unique identifier and authorization key
        image_uids:
          type: array
          items:
            type: string
          description: Array of authorized image UIDs
        allow_download:
          type: boolean
          description: Whether downloads are permitted with this token
          default: true
        allow_embed:
          type: boolean
          description: Whether embedding on external sites is allowed (false prevents hotlinking)
          default: false
        show_metadata:
          type: boolean
          description: Whether to include EXIF and metadata in responses
          default: true
        password:
          type: string
          nullable: true
          maxLength: 255
          description: Optional bcrypt hash of password (null if no password protection)
        description:
          type: string
          nullable: true
          maxLength: 500
          description: Optional description of this download link
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: When this token expires (null for no expiry)
        created_at:
          type: string
          format: date-time
          description: When this token was created
        updated_at:
          type: string
          format: date-time
          description: When this token was last updated
      required:
        [
          uid,
          image_uids,
          allow_download,
          allow_embed,
          show_metadata,
          created_at,
          updated_at,
        ]
...